pyproject.toml
[tool.poetry]
name = "desafio-google-sheets"
version = "0.1.0"
description = ""
authors = ["Tafarel Yan <tafarel.yan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^2.17.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
